requirements.txt
yt-dlp
openai-whisper
ffmpeg-python
gitpython
ffmpeg-python